pyproject.toml
[project]
name = "mcp-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "greenlet>=3.2.2",
    "langchain-google-genai>=2.1.4",
    "langchain-mcp-adapters>=0.1.1",
    "langgraph>=0.4.7",
    "playwright>=1.52.0",
    "pyee>=13.0.0",
    "python-dotenv>=1.1.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "filetype==1.2.0",
    "google-ai-generativelanguage==0.6.18",
    "google-api-core==2.24.2",
    "google-auth==2.40.2",
    "googleapis-common-protos==1.70.0",
    "greenlet==3.2.2",
    "grpcio==1.71.0",
    "grpcio-status==1.71.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "idna==3.10",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain-core==0.3.61",
    "langchain-google-genai==2.1.4",
    "langchain-mcp-adapters==0.1.1",
    "langgraph==0.4.7",
    "langgraph-checkpoint==2.0.26",
    "langgraph-prebuilt==0.2.1",
    "langgraph-sdk==0.1.70",
    "langsmith==0.3.42",
    "mcp==1.9.1",
    "orjson==3.10.18",
    "ormsgpack==1.10.0",
    "packaging==24.2",
    "playwright==1.52.0",
    "proto-plus==1.26.1",
    "protobuf==5.29.4",
    "pyasn1==0.6.1",
    "pyasn1_modules==0.4.2",
    "pydantic==2.11.5",
    "pydantic_core==2.33.2",
    "pydantic-settings==2.9.1",
    "pyee==13.0.0",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "PyYAML==6.0.2",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "rsa==4.9.1",
    "sniffio==1.3.1",
    "sse-starlette==2.3.5",
    "starlette==0.46.2",
    "tenacity==9.1.2",
    "typing_extensions==4.13.2",
    "typing-inspection==0.4.1",
    "urllib3==2.4.0",
    "uvicorn==0.34.2",
    "xxhash==3.5.0",
    "zstandard==0.23.0",
]
